Desktop/backend_django/requirements.txt
# Core Django
Django==5.0.1
djangorestframework==3.14.0
psycopg2-binary==2.9.9

# Authentication
django-allauth==0.57.0
dj-rest-auth==5.0.2

# File uploads and storage
Pillow==10.2.0

# Real-time WebSockets (install separately if needed)
# channels[daphne]==4.0.0
# channels-redis==4.2.0

# Environment variables
python-dotenv==1.0.1

# CORS
django-cors-headers==4.3.1

# Utilities
pytz==2024.1

# Production server
gunicorn==21.2.0
